+++ avivator/src/components/ChannelController.js
@@ -4,7 +4,6 @@ import Checkbox from '@material-ui/core/Checkbox';
 import Grid from '@material-ui/core/Grid';
 import Slider from '@material-ui/core/Slider';
 import Select from '@material-ui/core/Select';
-import Typography from '@material-ui/core/Typography';
 import CircularProgress from '@material-ui/core/CircularProgress';
 
 import { makeStyles } from '@material-ui/core/styles';
@@ -12,9 +11,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import ChannelOptions from './ChannelOptions';
 import { FILL_PIXEL_VALUE } from '../constants';
 
-const MIN_SLIDER_VALUE = 0;
-const MAX_SLIDER_VALUE = 65535;
-const COLORMAP_SLIDER_CHECKBOX_COLOR = [220, 220, 220];
+export const COLORMAP_SLIDER_CHECKBOX_COLOR = [220, 220, 220];
 
 const toRgb = (on, arr) => {
   const color = on ? COLORMAP_SLIDER_CHECKBOX_COLOR : arr;
@@ -31,16 +28,24 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
+function truncateDecimalNumber(value, maxLength) {
+  if (!value && value !== 0) return '';
+  const stringValue = value.toString();
+  return stringValue.length > maxLength
+    ? stringValue.substring(0, maxLength).replace(/\.$/, '')
+    : stringValue;
+}
+
 // If the channel is not on, display nothing.
 // If the channel has a not-undefined value, show it.
 // Otherwise, show a circular progress animation.
 const getPixelValueDisplay = (isOn, pixelValue, shouldShowPixelValue) => {
   if (!isOn || !shouldShowPixelValue) {
-    return <Typography> {FILL_PIXEL_VALUE} </Typography>;
+    return FILL_PIXEL_VALUE;
   }
   // Need to check if it's a number becaue 0 is falsy.
   if (pixelValue || typeof pixelValue === 'number') {
-    return <Typography> {pixelValue} </Typography>;
+    return truncateDecimalNumber(pixelValue, 7);
   }
   return <CircularProgress size="50%" />;
 };
@@ -55,7 +60,8 @@ function ChannelController({
   handleChange,
   pixelValue,
   shouldShowPixelValue,
-  disableOptions = false
+  disableOptions = false,
+  domain
 }) {
   const rgbColor = toRgb(colormapOn, colorValue);
   const classes = useStyles();
@@ -107,8 +113,9 @@ function ChannelController({
             onChange={(e, v) => handleChange('CHANGE_SLIDER', v)}
             valueLabelDisplay="auto"
             getAriaLabel={() => `${name}-${colorValue}-${sliderValue}`}
-            min={MIN_SLIDER_VALUE}
-            max={MAX_SLIDER_VALUE}
+            valueLabelFormat={v => truncateDecimalNumber(v, 5)}
+            min={domain[0]}
+            max={domain[1]}
             orientation="horizontal"
             style={{
               color: rgbColor,