+++ src/layers/ImageLayer.js
@@ -3,7 +3,7 @@ import { isWebGL2 } from '@luma.gl/core';
 
 import XRLayer from './XRLayer';
 import { padTileWithZeros } from '../loaders/utils';
-import { to32BitFloat } from './utils';
+import { to32BitFloat, onPointer } from './utils';
 
 const defaultProps = {
   pickable: true,
@@ -17,6 +17,7 @@ const defaultProps = {
   translate: { type: 'array', value: [0, 0], compare: true },
   scale: { type: 'number', value: 1, compare: true },
   boxSize: { type: 'number', value: 0, compare: true },
+  viewportId: { type: 'string', value: '', compare: true },
   loader: {
     type: 'object',
     value: {
@@ -25,7 +26,12 @@ const defaultProps = {
     },
     compare: true
   },
-  z: { type: 'number', value: 0, compare: true }
+  z: { type: 'number', value: 0, compare: true },
+  isLensOn: { type: 'boolean', value: false, compare: true },
+  lensSelection: { type: 'number', value: 0, compare: true },
+  lensRadius: { type: 'number', value: 100, compare: true },
+  lensBorderColor: { type: 'array', value: [255, 255, 255], compare: true },
+  lensBorderRadius: { type: 'number', value: 0.02, compare: true }
 };
 
 function scaleBounds({ width, height, translate, scale }) {
@@ -65,9 +71,20 @@ function padEven(data, width, height, boxSize) {
  * @param {Object} props.loader Loader to be used for fetching data.  It must implement/return `getRaster` and `dtype`.
  * @param {String} props.onHover Hook function from deck.gl to handle hover objects.
  * @param {String} props.boxSize If you want to pad an incoming tile to be a certain squared pixel size, pass the number here (only used by OverviewLayer/VivViewerLayer for now).
+ * @param {boolean} props.isLensOn Whether or not to use the lens.
+ * @param {number} props.lensSelection Numeric index of the channel to be focused on by the lens.
+ * @param {number} props.lensRadius Pixel radius of the lens (default: 100).
+ * @param {number} props.lensBorderColor RGB color of the border of the lens.
+ * @param {number} props.lensBorderRadius Percentage of the radius of the lens for a border (default 0.02).
  */
 export default class ImageLayer extends CompositeLayer {
   initializeState() {
+    this.state = {
+      unprojectLensBounds: [0, 0, 0, 0],
+      width: 0,
+      height: 0,
+      data: []
+    };
     const { loader, z, loaderSelection, boxSize } = this.props;
     loader.getRaster({ z, loaderSelection }).then(({ data, width, height }) => {
       this.setState(
@@ -79,6 +96,13 @@ export default class ImageLayer extends CompositeLayer {
         )
       );
     });
+    if (this.context.deck) {
+      this.context.deck.eventManager.on({
+        pointermove: () => onPointer(this),
+        pointerleave: () => onPointer(this),
+        wheel: () => onPointer(this)
+      });
+    }
   }
 
   updateState({ changeFlags, props, oldProps }) {
@@ -128,10 +152,14 @@ export default class ImageLayer extends CompositeLayer {
       z,
       domain,
       pickable,
+      isLensOn,
+      lensSelection,
+      lensBorderColor,
+      lensRadius,
       id
     } = this.props;
     const { dtype } = loader;
-    const { data, width, height } = this.state;
+    const { data, width, height, unprojectLensBounds } = this.state;
     if (!(width && height)) return null;
     const bounds = scaleBounds({
       width,
@@ -140,7 +168,7 @@ export default class ImageLayer extends CompositeLayer {
       scale
     });
     return new XRLayer({
-      channelData: Promise.resolve({ data, width, height }),
+      channelData: { data, width, height },
       pickable,
       bounds,
       sliderValues,
@@ -152,7 +180,12 @@ export default class ImageLayer extends CompositeLayer {
       opacity,
       visible,
       dtype,
-      colormap
+      colormap,
+      unprojectLensBounds,
+      isLensOn,
+      lensSelection,
+      lensBorderColor,
+      lensRadius
     });
   }
 }