+++ src/layers/MultiscaleImageLayer/MultiscaleImageLayer.js
@@ -3,11 +3,23 @@ import { isWebGL2 } from '@luma.gl/core';
 
 import MultiscaleImageLayerBase from './MultiscaleImageLayerBase';
 import ImageLayer from '../ImageLayer';
-import { to32BitFloat, getNearestPowerOf2 } from '../utils';
+import { to32BitFloat, getNearestPowerOf2, onPointer } from '../utils';
 
 const defaultProps = {
   pickable: true,
-  onHover: { type: 'function', value: null, compare: false }
+  onHover: { type: 'function', value: null, compare: false },
+  sliderValues: { type: 'array', value: [], compare: true },
+  colorValues: { type: 'array', value: [], compare: true },
+  channelIsOn: { type: 'array', value: [], compare: true },
+  opacity: { type: 'number', value: 1, compare: true },
+  colormap: { type: 'string', value: '', compare: true },
+  domain: { type: 'array', value: [], compare: true },
+  viewportId: { type: 'string', value: '', compare: true },
+  isLensOn: { type: 'boolean', value: false, compare: true },
+  lensSelection: { type: 'number', value: 0, compare: true },
+  lensRadius: { type: 'number', value: 100, compare: true },
+  lensBorderColor: { type: 'array', value: [255, 255, 255], compare: true },
+  lensBorderRadius: { type: 'number', value: 0.02, compare: true }
 };
 
 /**
@@ -25,9 +37,27 @@ const defaultProps = {
  * @param {String} props.id Unique identifier for this layer.
  * @param {String} props.onTileError Custom override for handle tile fetching errors.
  * @param {String} props.onHover Hook function from deck.gl to handle hover objects.
+ * @param {boolean} props.isLensOn Whether or not to use the lens.
+ * @param {number} props.lensSelection Numeric index of the channel to be focused on by the lens.
+ * @param {number} props.lensRadius Pixel radius of the lens (default: 100).
+ * @param {number} props.lensBorderColor RGB color of the border of the lens (default [255, 255, 255]).
+ * @param {number} props.lensBorderRadius Percentage of the radius of the lens for a border (default 0.02).
  */
 
 export default class MultiscaleImageLayer extends CompositeLayer {
+  initializeState() {
+    this.state = {
+      unprojectLensBounds: [0, 0, 0, 0]
+    };
+    if (this.context.deck) {
+      this.context.deck.eventManager.on({
+        pointermove: () => onPointer(this),
+        pointerleave: () => onPointer(this),
+        wheel: () => onPointer(this)
+      });
+    }
+  }
+
   renderLayers() {
     const {
       loader,
@@ -42,9 +72,14 @@ export default class MultiscaleImageLayer extends CompositeLayer {
       onTileError,
       onHover,
       pickable,
-      id
+      id,
+      isLensOn,
+      lensSelection,
+      lensBorderColor,
+      lensBorderRadius
     } = this.props;
     const { tileSize, numLevels, dtype } = loader;
+    const { unprojectLensBounds } = this.state;
     const noWebGl2 = !isWebGL2(this.context.gl);
     const getTileData = async ({ x, y, z }) => {
       const tile = await loader.getTile({
@@ -96,7 +131,12 @@ export default class MultiscaleImageLayer extends CompositeLayer {
       colormap,
       viewportId,
       onHover,
-      pickable
+      pickable,
+      unprojectLensBounds,
+      isLensOn,
+      lensSelection,
+      lensBorderColor,
+      lensBorderRadius
     });
     // This gives us a background image and also solves the current
     // minZoom funny business.  We don't use it for the background if we have an opacity
@@ -116,7 +156,7 @@ export default class MultiscaleImageLayer extends CompositeLayer {
           (-numLevels > this.context.viewport.zoom &&
             (!viewportId || this.context.viewport.id === viewportId)),
         z: numLevels - 1,
-        pickable: true,
+        pickable: false,
         onHover,
         boxSize: getNearestPowerOf2(lowResWidth, lowResHeight)
       });