+++ src/layers/MultiscaleImageLayer/MultiscaleImageLayerBase.js
@@ -16,7 +16,13 @@ const defaultProps = {
   colormap: { type: 'string', value: '', compare: true },
   dtype: { type: 'string', value: '<u2', compare: true },
   domain: { type: 'array', value: [], compare: true },
-  viewportId: { type: 'string', value: '', compare: true }
+  viewportId: { type: 'string', value: '', compare: true },
+  unprojectLensBounds: { type: 'array', value: [0, 0, 0, 0], compare: true },
+  isLensOn: { type: 'boolean', value: false, compare: true },
+  lensSelection: { type: 'number', value: 0, compare: true },
+  lensRadius: { type: 'number', value: 100, compare: true },
+  lensBorderColor: { type: 'array', value: [255, 255, 255], compare: true },
+  lensBorderRadius: { type: 'number', value: 0.02, compare: true }
 };
 
 /**