+++ src/layers/MultiscaleImageLayer/utils.js
@@ -6,7 +6,10 @@ export function range(len) {
 
 export function renderSubLayers(props) {
   const {
-    bbox: { left, top, right, bottom }
+    bbox: { left, top, right, bottom },
+    x,
+    y,
+    z
   } = props.tile;
   const {
     colorValues,
@@ -19,14 +22,17 @@ export function renderSubLayers(props) {
     dtype,
     id,
     onHover,
-    pickable
+    pickable,
+    unprojectLensBounds,
+    isLensOn,
+    lensSelection
   } = props;
   // Only render in positive coorinate system
-  if ([left, top, right, bottom].some(v => v < 0)) {
+  if ([left, bottom, right, top].some(v => v < 0)) {
     return null;
   }
   const xrl = new XRLayer({
-    id: `XRLayer-${left}-${top}-${right}-${bottom}-${id}`,
+    id: `XRLayer-${left}-${bottom}-${right}-${top}-${id}`,
     bounds: [left, bottom, right, top],
     channelData: data,
     pickable,
@@ -41,7 +47,11 @@ export function renderSubLayers(props) {
     visible,
     dtype,
     colormap,
-    onHover
+    onHover,
+    unprojectLensBounds,
+    isLensOn,
+    lensSelection,
+    tileId: { x, y, z }
   });
   return xrl;
 }